requirements.txt
asgiref==3.8.1
awsebcli==3.20.10
botocore==1.31.85
cement==2.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.3
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
gunicorn==22.0.0
idna==3.7
jmespath==1.0.1
packaging==24.1
pathspec==0.10.1
pypiwin32==223
python-dateutil==2.9.0.post0
pywin32==306
PyYAML==6.0.1
requests==2.32.3
semantic-version==2.8.5
setuptools==71.0.4
six==1.16.0
sqlparse==0.5.1
termcolor==1.1.0
tzdata==2024.1
urllib3==1.26.19
wcwidth==0.1.9
